fix(activerecord): move the Migration AR-config registry to an import-free leaf
Entering the adapter graph through `SchemaStatements` evaluates
`base.ts`'s module body while `migration.ts` is still mid-evaluation
(schema-statements → … → migration.ts → schema-migration.ts → base.ts),
so `registerMigrationArConfig` ran before `migration.ts` reached its
own `let _arConfig` declaration and crashed with
`ReferenceError: Cannot access '_arConfig' before initialization`.
This broke the `scripts/test-deps/adapter-graph-import-tdz.test.ts`
regression guard, failing the Unit Tests job on main.

Move `MigrationArConfig`, `registerMigrationArConfig` and its readers
into `migration/ar-config-source.ts`, an import-free leaf module —
the same shape `connection-adapters/abstract/table-name-options.ts`
already uses for Base's other self-registration. A leaf with no imports
is always fully evaluated before any importer's body runs, so the
registration can never land in a temporal dead zone.
